Refactor #delete method
Delete method WAS working with leaf nodes and nodes with
a single child. It would even replace a node with two children
with its appropriate successor.

However, the left side of the deleted node would not restructure
properly and nodes were lost. Read through some examples in Python
and C++, and realized I would not be able to perform the proper
recursive method with my current code.

Refactored with similar structure to examples and was able to
eliminate three helper functions I had built previously. Code is
much cleaner now.